+++ app/app/handlers/forms/moderator/add_view_job.py
@@ -10,6 +10,7 @@ from dependency_injector.wiring import inject, Provide
 from app.services.application import ApplicationService
 from app.core.container import Container
 from app.models.application import Application
+from loguru import logger
 
 
 async def get_sub_object(message: types.Message, state: FSMContext):
@@ -35,9 +36,13 @@ async def get_sort(message: types.Message, state: FSMContext):
     await state.set_state(AddViewWork.subsystems)
 
 
-async def get_subsystems(query: types.CallbackQuery, state: FSMContext):
+@inject
+async def get_subsystems(query: types.CallbackQuery, state: FSMContext,
+                         application: ApplicationService = Provide[Container.application_service]):
     new_kb = kb.accept().add(kb.exit_button)
     data = await state.get_data()
+    logger.info(data)
+    logger.info(len(data))
     if query.data != 'accept' and 'field_four' not in data:
         await state.update_data(field_four=query.data)
         data = await state.get_data()
@@ -48,15 +53,24 @@ async def get_subsystems(query: types.CallbackQuery, state: FSMContext):
         await state.update_data(field_four=data['field_four'] + ', ' + query.data)
         new_data = await state.get_data()
         await bot.edit_message_text(new_data['field_four'],
-                                    query.message.chat.id, new_data['message_id'], reply_markup=new_kb)
+                                    query.message.chat.id, data['message_id'], reply_markup=new_kb)
     else:
-        await bot.delete_message(
-            query.message.chat.id, query.message.message_id)
-        await get_data.send_data(query=query, state=state)
-        new_kb = kb.sure().add(kb.exit_button)
-        await query.message.answer(const.SURE,
-                                   reply_markup=new_kb)
-        await state.set_state(AddViewWork.sure)
+        if 'admin' not in data:
+            await bot.delete_message(
+                query.message.chat.id, query.message.message_id)
+            await get_data.send_data(query=query, state=state)
+            new_kb = kb.sure().add(kb.exit_button)
+            await query.message.answer(const.SURE, reply_markup=new_kb)
+            await state.set_state(AddViewWork.sure)
+        else:
+            black_list = {'admin'}
+            new_data = {key: val for key, val in data.items() if key not in black_list}
+            unused = ['field_five', 'field_six', 'field_seven', 'field_eight', 'field_nine']
+            for i in unused:
+                new_data[i] = None
+            await application.update(data['admin'], obj_in=new_data)
+            await query.message.answer(const.CHANGE_SUCCESS)
+            await state.finish()
     await query.answer()
 
 
@@ -76,13 +90,18 @@ async def correct(query: types.CallbackQuery, state: FSMContext):
                                    reply_markup=new_kb)
         await state.set_state(AddViewWork.edit)
     elif query.data == '3':
-        await bot.delete_message(
-            query.message.chat.id, query.message.message_id)
-        await state.update_data(change='request_type')
-        new_kb = kb.main_kb().add(kb.exit_button)
-        await query.message.answer(R_TYPE,
-                                   reply_markup=new_kb)
-        await state.set_state(BaseStates.request_type)
+        data = await state.get_data()
+        if 'admin' in data:
+            await bot.delete_message(query.message.chat.id,
+                                     query.message.message_id)
+            await query.message.answer(text=FIO, reply_markup=kb.exit_kb())
+            await state.set_state(BaseStates.fio)
+        else:
+            await bot.delete_message(
+                query.message.chat.id, query.message.message_id)
+            await state.finish()
+            await query.message.answer(text=FIO, reply_markup=kb.exit_kb())
+            await state.set_state(BaseStates.fio)
     elif query.data == '4':
         await bot.delete_message(
             query.message.chat.id, query.message.message_id)
